Donation check

+++ donor/views.py
@@ -10,6 +10,7 @@ from django.core.mail import EmailMessage
 from django.conf import settings
 from django.template.loader import render_to_string
 from django.core.paginator import Paginator
+from django.utils import timezone
 
 # Create your views here.
 
@@ -62,40 +63,52 @@ def donation(request, request_id):
     if request.method == 'POST':
         try:
             user_profile = request.user.userprofile
-            user_profile.donor_donationcount += 1
-            user_profile.save()
-
-
             recipient_obj = get_object_or_404(recipient, pk=request_id)
-            recipient_name = recipient_obj.recipientname
-            recipient_email = recipient_obj.recipientemail
 
-            recipient_email_subject = "Request Confirmation"
-            recipient_message = render_to_string('recipientemail.html', {
-                'donor_name': request.user.first_name,
-                'donor_phone': request.user.userprofile.phone,
-                'donor_email': request.user.email,
-                'donor_location': request.user.userprofile.location,
-                'donor_bloodgroup': request.user.userprofile.bloodgroup,
-                'donor_age': request.user.userprofile.age,
-                'donor_gender': request.user.userprofile.gender,
-                'recipient_name' : recipient_name,
-                'domain': '127.0.0.1:8000',
-            })
+            if request.user.userprofile.bloodgroup == recipient_obj.recipientblood:
 
-            recipient_email_message = EmailMessage(
-                recipient_email_subject,
-                recipient_message,
-                settings.EMAIL_HOST_USER,
-                [recipient_email]
-            )
-            recipient_email_message.send()
+                if timezone.now() < user_profile.next_eligible_date:
+                    messages.warning(request, 'YOU NEED TO WAIT FOR 12 WEEKS AFTER YOUR LAST DONATION.')
+                    return redirect('profile')
+                
+                else:
+                    user_profile.donor_donationcount += 1
+                    user_profile.last_donation = timezone.now()
+                    user_profile.save()
 
 
-            recipient_obj.delete()
+                recipient_name = recipient_obj.recipientname
+                recipient_email = recipient_obj.recipientemail
 
-            messages.success(request, 'THANK YOU FOR YOUR DONATION')
-            return redirect('profile')
+                recipient_email_subject = "Request Confirmation"
+                recipient_message = render_to_string('recipientemail.html', {
+                    'donor_name': request.user.first_name,
+                    'donor_phone': request.user.userprofile.phone,
+                    'donor_email': request.user.email,
+                    'donor_location': request.user.userprofile.location,
+                    'donor_bloodgroup': request.user.userprofile.bloodgroup,
+                    'donor_age': request.user.userprofile.age,
+                    'donor_gender': request.user.userprofile.gender,
+                    'recipient_name' : recipient_name,
+                    'domain': '127.0.0.1:8000',
+                })
+
+                recipient_email_message = EmailMessage(
+                    recipient_email_subject,
+                    recipient_message,
+                    settings.EMAIL_HOST_USER,
+                    [recipient_email]
+                )
+                recipient_email_message.send()
+
+
+                recipient_obj.delete()
+
+                messages.success(request, 'THANK YOU FOR YOUR DONATION')
+                return redirect('profile')
+            else:
+                messages.error(request, 'BLOOD GROUP NOT MATCHED.')
+                return redirect('/donor/donation-request/')
         
         except userprofile.DoesNotExist:
             messages.error(request, 'REGISTER AS DONOR TO DONATE BLOOD.')
